+++ TestTask/Extensions/ListExtension.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace TestTask
+{
+    public static class ListExtension
+    {
+
+        public static int RemoveAll<T>(this IList<T> list, Predicate<T> match)
+        {
+            int count = 0;
+
+            for (int i = list.Count - 1; i >= 0; i--)
+            {
+                if (match(list[i]))
+                {
+                    ++count;
+                    list.RemoveAt(i);
+                }
+            }
+            return count;
+        }
+
+    }
+}
