+++ TestTask/Extensions/StringExtension.cs
