--- TestTask/IReadOnlyStream.cs
@@ -1,19 +1,0 @@
-﻿using System;
-
-namespace TestTask
-{
-    /// <summary>
-    /// Интерфейс для работы с файлом в сильно урезаном виде.
-    /// Умеет всего 2 вещи: прочитать символ, и перемотать стрим на начало.
-    /// </summary>
-    public interface IReadOnlyStream : IDisposable
-    {
-        char ReadNextChar();
-
-        void Close(); 
-
-        void ResetPositionToStart();
-
-        bool IsEof { get; }
-    }
-}
