+++ TestTask/TestTask.csproj
@@ -43,13 +43,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CharType.cs" />
-    <Compile Include="IReadOnlyStream.cs" />
-    <Compile Include="LetterStats.cs" />
+    <Compile Include="Enums\CharType.cs" />
+    <Compile Include="Extensions\ListExtension.cs" />
+    <Compile Include="Abstraction\IReadOnlyStream.cs" />
+    <Compile Include="Models\LetterStats.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ReadOnlyStream.cs" />
-    <Compile Include="StringExtension.cs" />
+    <Compile Include="Utilities\ReadOnlyStream.cs" />
+    <Compile Include="Extensions\StringExtension.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
